Move NoteFilename and NoteDirPath from todo.go to note.go
